V4L/DVB (7352): tuner-simple: enable digital tuning support for Philips FMD1216ME
Enable digital tuning support within tuner-simple. This will allow for a
single tuner module to manage the hardware, without having dvb-pll loaded.

+++ include/media/tuner-types.h
@@ -119,7 +119,7 @@ struct tunertype {
 
 	u16 min;
 	u16 max;
-	u16 stepsize;
+	u32 stepsize;
 
 	u8 *initdata;
 	u8 *sleepdata;
